+++ Native/Native.cpp
@@ -10,7 +10,7 @@ NATIVE_API int Add(int a, int b)
     return a + b;
 }
 
-NATIVE_API BYTE* NewBuffer(__int64 cb) {
+NATIVE_API BYTE* NewBuffer(size_t cb) {
     return new BYTE[cb];
 }
 
@@ -18,7 +18,7 @@ NATIVE_API void DeleteBuffer(BYTE* buffer) {
     delete [] buffer;
 }
 
-NATIVE_API BYTE* MallocBuffer(__int64 cb) {
+NATIVE_API BYTE* MallocBuffer(size_t cb) {
     return (BYTE*)malloc(cb);
 }
