+++ Native/Native.h
@@ -11,7 +11,7 @@
 #endif
 
 NATIVE_API int Add(int a, int b);
-NATIVE_API BYTE* NewBuffer(__int64 cb);
+NATIVE_API BYTE* NewBuffer(size_t cb);
 NATIVE_API void DeleteBuffer(BYTE* buffer);
-NATIVE_API BYTE* MallocBuffer(__int64 cb);
+NATIVE_API BYTE* MallocBuffer(size_t cb);
 NATIVE_API void FreeBuffer(BYTE* buffer);
